+++ test_characteristic.py
@@ -123,6 +123,45 @@ class TestAttribute(object):
             " init_aliaser=None)>"
         ).format("type" if PY2 else "class") == repr(a)
 
+    def test_eq_different_types(self):
+        """
+        Comparing Attribute with something else returns NotImplemented.
+        """
+        assert NotImplemented == Attribute(name="name").__eq__(None)
+
+    def test_eq_equal(self):
+        """
+        Equal Attributes are detected equal.
+        """
+        kw = {
+            "name": "name",
+            "exclude_from_cmp": True,
+            "exclude_from_init": False,
+            "exclude_from_repr": True,
+            "exclude_from_immutable": False,
+            "default_value": 42,
+            "instance_of": int,
+        }
+        assert Attribute(**kw) == Attribute(**kw)
+
+    def test_eq_unequal(self):
+        """
+        Equal Attributes are detected equal.
+        """
+        kw = {
+            "name": "name",
+            "exclude_from_cmp": True,
+            "exclude_from_init": False,
+            "exclude_from_repr": True,
+            "exclude_from_immutable": False,
+            "default_value": 42,
+            "instance_of": int,
+        }
+        for arg in kw.keys():
+            kw_mutated = dict(**kw)
+            kw_mutated[arg] = "mutated"
+            assert Attribute(**kw) != Attribute(**kw_mutated)
+
 
 @with_cmp(["a", "b"])
 class CmpC(object):
@@ -562,6 +601,45 @@ class TestAttributes(object):
             pass
 
         assert repr(C(a=1)).startswith("<test_characteristic.")
+
+    def test_store_attributes(self):
+        """
+        store_attributes is called on the class to store the attributes that
+        were passed in.
+        """
+        attrs = [Attribute("a"), Attribute("b")]
+
+        @attributes(
+            attrs, store_attributes=lambda cls, a: setattr(cls, "foo", a)
+        )
+        class C(object):
+            pass
+
+        assert C.foo == attrs
+
+    def test_store_attributes_stores_Attributes(self):
+        """
+        The attributes passed to store_attributes are always instances of
+        Attribute, even if they were simple strings when provided.
+        """
+        @attributes(["a", "b"])
+        class C(object):
+            pass
+
+        assert C.characteristic_attributes == [Attribute("a"), Attribute("b")]
+
+    def test_store_attributes_defaults_to_characteristic_attributes(self):
+        """
+        By default, store_attributes stores the attributes in
+        `characteristic_attributes` on the class.
+        """
+        attrs = [Attribute("a")]
+
+        @attributes(attrs)
+        class C(object):
+            pass
+
+        assert C.characteristic_attributes == attrs
 
     def test_optimizes(self):
         """